Extract IssueFinder to its own file
